move tests to scala 2.11+ to avoid 6260 in 2.10
